Altered the file structure of the models and the controllers
