--- app/controllers/delete-comment-controller.js
@@ -1,12 +1,0 @@
-const { removeComment } = require("../models");
-
-exports.deleteComment = (request, response, next) => {
-  const { comment_id } = request.params;
-  removeComment(comment_id)
-    .then(() => {
-      response.sendStatus(204);
-    })
-    .catch((err) => {
-      next(err);
-    });
-};
