--- app/controllers/get-review-controller.js
@@ -1,12 +1,0 @@
-const { fetchReview } = require("../models");
-
-exports.getReview = (request, response, next) => {
-  const { review_id } = request.params;
-  fetchReview(review_id)
-    .then(([review]) => {
-      response.status(200).send({ review });
-    })
-    .catch((err) => {
-      next(err);
-    });
-};
