--- app/models/delete-comment-controller.js
@@ -1,16 +1,0 @@
-const db = require("../../db/connection");
-
-exports.removeComment = (comment_id) => {
-  return db
-    .query(`DELETE FROM comments WHERE comment_id = $1 RETURNING *`, [
-      comment_id,
-    ])
-    .then(({ rows }) => {
-      if (!rows.length) {
-        return Promise.reject({
-          status: 404,
-          message: "Requested Item Not Found Within the Database",
-        });
-      }
-    });
-};
