+++ app/models/index.js
@@ -1,11 +1,15 @@
 const { fetchCategories } = require("./get-categories-model");
-const { fetchReview } = require("./get-review-model");
-const { changeVotesProperty } = require("./patch-review-model");
+const {
+  fetchReview,
+  fetchAllReviews,
+  changeVotesProperty,
+} = require("./review-model");
 const { fetchUsers } = require("./get-users-model");
-const { addNewComment } = require("./post-comment-model");
-const { fetchAllReviews } = require("./get-all-reviews");
-const { removeComment } = require("./delete-comment-controller");
-const { fetchComments } = require("./get-comments-model");
+const {
+  fetchComments,
+  addNewComment,
+  removeComment,
+} = require("./comments-model");
 
 module.exports = {
   fetchCategories,
