+++ app/models/review-model.js
@@ -1,5 +1,21 @@
 const db = require("../../db/connection");
 
+exports.fetchReview = (review_id) => {
+  return db
+    .query(
+      "SELECT reviews.*, COUNT (comment_id) ::int AS comment_count FROM reviews LEFT JOIN comments ON comments.review_id = reviews.review_id WHERE reviews.review_id = $1 GROUP BY reviews.review_id;",
+      [review_id]
+    )
+    .then(({ rows }) => {
+      if (!rows.length) {
+        return Promise.reject({
+          status: 404,
+          message: "Requested Item Not Found Within the Database",
+        });
+      } else return rows;
+    });
+};
+
 exports.fetchAllReviews = (
   sort_by = "created_at",
   order = "DESC",
@@ -48,4 +64,20 @@ exports.fetchAllReviews = (
       });
     } else return rows;
   });
+};
+
+exports.changeVotesProperty = (review_id, inc_votes) => {
+  return db
+    .query(
+      `UPDATE reviews SET votes = votes + $1 WHERE review_id = $2 RETURNING *;`,
+      [inc_votes, review_id]
+    )
+    .then(({ rows }) => {
+      if (!rows.length) {
+        return Promise.reject({
+          status: 404,
+          message: "Requested Item Not Found Within the Database",
+        });
+      } else return rows[0];
+    });
 };
